pipeline/requirements.txt
lightrag-hku>=1.3.0
sentence-transformers>=3.0.0
openai>=1.50.0
qdrant-client>=1.12.0
neo4j>=5.0.0
python-frontmatter>=1.1.0

# MCP server (local only, not used in GitHub Actions)
mcp>=1.0.0
python-dotenv>=1.0.0
